world position support
given we added parenting, this change adds utility methods for
computing the world position of an entity. entitiy positions
are stored in local space. fixed a dumb bug where i was rounding
floating positions in the translation gizmo code. this lead to eniies
awkwardly snapping along an axis that it shouldn't be moving along